% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_boxcox.R
\name{apply_boxcox}
\alias{apply_boxcox}
\title{Apply the boxcox transformation to a linear model}
\usage{
apply_boxcox(model, digits = 1, transformed = TRUE, plotit = FALSE)
}
\arguments{
\item{model}{an object of type \code{lm} to be transformed}

\item{digits}{desired number of digits for calculated \code{lambda} in the transformation.}

\item{transformed}{logical paramter. If is TRUE the function returns the vector of transformed response.}

\item{plotit}{logical parameter. If is TRUE a graph of log-likelihood is displayed}
}
\value{
The value of \code{lambda}, for apply the transformation.

The vector of \code{transformed response}.
}
\description{
This function returns the transformed response variable when apply the boxcox transformation to an specific linear model
}
\details{
For more information about boxcox transformation and when it is convenient to use \href{https://en.wikipedia.org/wiki/Power_transform}{see this link}
}
\examples{
y <- runif(10,0,20)
x <- rnorm(10)
apply_boxcox(lm(y~x))
apply_boxcox(lm(y~x), digits = 2)
}
