% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bctrans.R
\name{bctrans}
\alias{bctrans}
\title{Box-Cox Transformation}
\usage{
bctrans(y, lambda, plot_it = TRUE)
}
\arguments{
\item{y}{Vector to be transformed.}

\item{lambda}{Power to raise \code{y} by.}

\item{plot_it}{Display a plot of \code{y} vs the output? Use logical.}
}
\value{
The vector \code{y}, Box-Cox transformed to a power of \code{lambda}.
}
\description{
This function applies Box-Cox transformation to a numerical vector.
}
