% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bctrans_inv.R
\name{bctrans_inv}
\alias{bctrans_inv}
\title{Inverse Box-Cox Transformation}
\usage{
bctrans_inv(z, lambda, plot_it = TRUE)
}
\arguments{
\item{z}{Vector to be inverse-transformed.}

\item{lambda}{Power to raise inverse \code{z} by.}

\item{plot_it}{Display a plot of \code{z} vs the output? Use logical.}
}
\value{
The vector \code{z}, Inversely Box-Cox transformed at a power of \code{lambda}.
}
\description{
This function applies Inverse Box-Cox transformation to a numerical vector.
}
