% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{box.cox}
\alias{box.cox}
\title{Apply box-cox transformation to smoothen non-linearity of data}
\usage{
box.cox(x, lambda)
}
\arguments{
\item{x}{is the vector(positive values) to apply box-cox transformation}

\item{lambda}{is the power parameter in box-cox transformation}
}
\value{
\itemize{
   \item if lambda = 0, ln(x)
   \item for all other lambda, (x^lambda-1)/lambda
}
}
\description{
Apply box-cox transformation to smoothen non-linearity of data
}
