% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrt.R
\name{cbrt}
\alias{cbrt}
\title{Apply cube root}
\usage{
cbrt(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to calculate cube root}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is a cube root of \code{x}.
}
\description{
Apply cube root
}
\examples{
cbrt(27)
}
