% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv-boxcox-matrix-of-val.R
\name{inv.boxcox.tranform.mat}
\alias{inv.boxcox.tranform.mat}
\title{Inverse Box-Cox Transform Matrix of Values Function}
\usage{
inv.boxcox.tranform.mat(x, lambda = 0)
}
\arguments{
\item{x}{vector that is to be transformed under an inverse box-cox function.}

\item{lambda}{vector of values that indicate the strength of the inverse box-cox transformation.}
}
\value{
a matrix of the `x` values that underwent an inverse box-cox transformation for different lambdas. The entries of a column correspond to a particular lambda value. Hence, the columns are named to reflect what lambda value was used.
}
\description{
This function is to be used in an inverse boxcox.transform function to return a matrix of values that
underwent an inverse box-cox transform when lambda is a vector.
Each column of the matrix corresponds to a different lambda (and is named according to the lambda value).
I'm making it internal, because I don't want users to use it.
Rather, I want to use it in the inv.boxcox.transform function.
}
