% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.boxcox.transform.R
\name{inv.boxcox.transform}
\alias{inv.boxcox.transform}
\title{Inverse Box-Cox Transform Function}
\usage{
inv.boxcox.transform(x, lambda = 0, makepos = 0)
}
\arguments{
\item{x}{vector that is to be transformed under the inverse box-cox function. All x must be positive.}

\item{lambda}{strength of the transformation. Default is 0. May be a single value or a vector for this function.}

\item{makepos}{the constant value that should be added to each `x`  to make all x positive. Default is 0.}
}
\value{
A vector of the values that have been transformed. If lambda is a 0 value, returns the `exp(x)`. If lambda is non-0 value, returns `(lambda*x + 1)^(1/lambda)`.
}
\description{
Given a variable, the inverse Box-Cox power transformation is obtained.
}
