% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function.R
\name{lm_plot}
\alias{lm_plot}
\title{Display linear model in plot}
\usage{
lm_plot(x, y, mydata)
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{mydata}{the dataframe to be analyzed}
}
\value{
A table and qplot
}
\description{
This function prints the results with simple linear regression model
in the form os a table or regression coefficients and qplot
}
