% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logarithm.R
\name{log3}
\alias{log3}
\alias{log4}
\alias{log5}
\title{Apply a useful logarithm}
\usage{
log3(x)

log4(x)

log5(x)
}
\arguments{
\item{x}{The vector to be log3, log4 or log5}
}
\value{
A vector that is the:
\itemize{
     \item log base 3 (for \code{log3})
     \item log base 4 (for \code{log4})
     \item log base 5 (for \code{log5})
}
 of \code{x}.
}
\description{
Used to having the basics? log2 and log10?
Now, you have other bases: log3, log4 and log5
}
\examples{
log3(tenvec)
log4(9)
log5(1:5)
}
