% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_boxcox.R
\name{need_boxcox}
\alias{need_boxcox}
\title{Does your linear model need the Boxcox transformation?}
\usage{
need_boxcox(model, conf = 0.95, plotit = TRUE)
}
\arguments{
\item{model}{an object of type \code{lm} to be tested}

\item{conf}{level of confidence for test (shoulb be between 0 and 1)}

\item{plotit}{logical parameter. If is TRUE a graph of test is displayed}
}
\value{
\code{suggestion} gives you the suggestion about if transformation is needed or not.
}
\description{
This function analyzes the residuals of your model and suggest you to apply or not the Boxcox transformation
}
\examples{
y <- runif(10,0,20)
x <- rnorm(10)
need_boxcox(lm(y~x))
}
