% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Convenient Computation of Powers}
\description{
Are you tired of using the power operator, \code{^} or \code{**} in R? Do you want an easy way to take roots or reciprocals?
This package will call functions that apply common powers or roots
to your vectors.
}
\note{
This package isn't actually meant to be serious. It's just for
teaching purposes.
This package was adapted by JasminLib for hw07 of Stat547.
}
\author{
Vincenzo Coia

Jasmine Lib
}
