% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\title{#' Apply a useful root}
\usage{
root(x, y, plot_it = FALSE)
}
\arguments{
\item{x}{the vector for which you want the root}

\item{y}{the integer root that you want to take (ie square root = 2, cube root = 3, etc)}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item root (for \code{root})
}
 of \code{x}.
}
\description{
#' Apply a useful root
}
