% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox-matrix-of-val.R
\docType{data}
\name{values_mat}
\alias{values_mat}
\title{Box-Cox Transform Matrix of Values Function}
\format{An object of class \code{matrix} with 5 rows and 2 columns.}
\usage{
values_mat
}
\arguments{
\item{x}{vector that is to be transformed under the box-cox function.}

\item{lambda}{vector of values that indicate the strength of the box-cox transformation}
}
\value{
a matrix of the `x` values that underwent box-cox transformation for different lambdas. The entries of a column correspond to a particular lambda value. Hence, the columns are named to reflect what lambda value was used.
}
\description{
This function is to be used in the boxcox.transform function to return a matrix of values that
underwent a box-cox transform when lambda is a vector.
Each column of the matrix corresponds to a different lambda (and is named according to the lambda value).
I'm making it internal, because I don't want users to use it.
Rather, I want to use it in the boxcoxtransform function.
}
\keyword{datasets}
