% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\title{Take the root of a numeric}
\usage{
root(x, y = 2, plot_it = FALSE)
}
\arguments{
\item{x}{the vector or numeric input for which you want to take a root}

\item{y}{the integer (or non-integer, as long as it's numeric) root that you want to take (ie square root = 2, cube root = 3, etc)}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item root of the first input (\code{x}.)
     \item will default to square root, though specific root can be specified in second argument with a numeric input such as y = 3, 4, 5, 6, etc..
}
}
\description{
The Root function will take the root (square, cube, etc, as specified) of a numeric input.
}
\details{
This function was developped as part of homework for STAT547 class. hopefully it is of some use to you!
For more details see vignette
}
\examples{
root(27,3)
root(c(0.01,1,4,9))
root(c(TRUE, FALSE, TRUE), 5)

}
